+++ agent/plan_runner.go
@@ -2,21 +2,18 @@ package agent
 
 import (
 	"github.com/go-logr/logr"
+	"github.com/leg100/go-tfe"
 	"github.com/leg100/ots"
 )
 
-func mockNewPlanRunnerFn(run *ots.Run,
-	cvs ots.ConfigurationVersionService,
-	svs ots.StateVersionService,
-	rs ots.RunService,
-	log logr.Logger) *ots.Runner {
+type NewPlanRunnerFn func(
+	*ots.Run,
+	ots.ConfigurationVersionService,
+	ots.StateVersionService,
+	ots.RunService,
+	logr.Logger) *ots.Runner
 
-	return ots.NewRunner(
-		[]ots.Step{},
-	)
-}
-
-func mockNewPlanRunnerFnWithError(run *ots.Run,
+func NewPlanRunner(run *ots.Run,
 	cvs ots.ConfigurationVersionService,
 	svs ots.StateVersionService,
 	rs ots.RunService,
@@ -24,7 +21,14 @@ func mockNewPlanRunnerFnWithError(run *ots.Run,
 
 	return ots.NewRunner(
 		[]ots.Step{
-			ots.NewCommandStep("/bin/false"),
+			DownloadConfigStep(run, cvs),
+			DeleteBackendStep,
+			DownloadStateStep(run, svs, log),
+			UpdatePlanStatusStep(run, rs, tfe.PlanRunning),
+			InitStep,
+			PlanStep,
+			JSONPlanStep,
+			FinishPlanStep(run, rs, log),
 		},
 	)
 }
